[maybe] Stop limiting many commands to only open topics
Previously many commands only offered open topics as completion
candidates and the user had to use a prefix argument to also be
offered closed topics.

This turned out to be weird in practice and outright annoying for
commands that are supposed to use the topic at point without even
letting the user pick another topic.  In such cases the user was
prompted and the topic at point wasn't even offered as a choice.

A single command (`forge-merge' which merges using the forges API)
now only ever offers open pull-requests.

A single command (`forge-branch-pullreq') continues to require the use
of a prefix argument to allow selecting a closed pull-request.  It now
takes care of handling the prefix-argument itself, and the respective
functionality can be removed from `forge-read-pullreq',
`forge-read-issue' and `forge-read-topic'.

We started offering only open topics by default in [1: 20d0dc8d] (#42)
but back then it was still possible to select the topic at point even
when that is closed and no prefix argument is used.  This special case
broke somewhere along the way.

IMO it would be best to let the user toggle between only open or all
topics when completion is already in progress.  Such a feature would
be highly dependent on the used completion framework and is left for
another day.

1: 2019-01-20 20d0dc8dbf33f6c9e897728750f3e0f91010b7d4
   Only offer open topics as completion candidates by default